+++ src/base/lemma/lib/graph-fw-v3-class.js
@@ -6,7 +6,7 @@
 // These creators, however dependent via createsLowTireGlobalCSS.
 
 ( function() {
-    var { $$, nsmethods, haz, globalCss, stdMod, }
+    var { $$, sDomF, nsmethods, haz, globalCss, stdMod, }
         = window.b$l.apptree({ stdModExportList : { createsGraphFW_class, }, });
     var GLOBAL_CSS_APPENDED = false;
     return;
@@ -20,13 +20,10 @@
     function createsGraphFW_class({
         graphFW,
         digramParentDom$,
-        doSetColorThreadArray,
-        setsGraphContainerAttributes,
-        setsGraphAxes,
-        plotLabels_2_plotsPars,
+        customXLegend,
     }){
         var graphFW__self = graphFW;
-        var colorThreadArray = graphFW__self.colorThreadArray = doSetColorThreadArray();
+        var colorThreadArray = graphFW__self.colorThreadArray = setColorThreadArray();
 
         //===========================================
         // //\\ fills wrap-object
@@ -36,7 +33,7 @@
         graphFW__self.showPHGraph = showPHGraph;
         ///creates fw-dom-container
         let {container$, graph_dimX, graph_dimY} =
-            setsGraphContainerAttributes( digramParentDom$ );
+            setGraphContainerAttributes( digramParentDom$ );
         ///creates low tier api
         graphFW__self.fw = nsmethods.createsGraphFramework({
             parent : container$,
@@ -71,7 +68,10 @@
             //first array mast be enabled
             let graphArrayMask = haz( graphFW__self, 'graphArrayMask' );
 
-            var { yColor, xColor, axisYLegend, axisXLegend, } = setsGraphAxes();
+			var { legendText, legendX } = customXLegend ? 
+				customXLegend() :
+				{ legendText: 'Distance from force (SP)', legendX : -560 };
+            var { textColor, textColor, axisYLegend, axisXLegend, } = setGraphAxes(graphFW, legendText, legendX);
             //==================================================
             // //\\ calls api
             // //\\ calls low tier api
@@ -85,8 +85,8 @@
                 style : {
                    //'stroke-width' : 2, //destroys tp-machine
                 },
-                axisX : graphAxisX( xColor ),
-                axisY : graphAxisY( yColor ),
+                axisX : graphAxisX( textColor ),
+                axisY : graphAxisY( textColor ),
                 drawDecimalY,
                 drawDecimalX,
                 doSideAxes : true,
@@ -98,7 +98,6 @@
                 axisYLegend,
                 axisXLegend,
                 plotsCount_overrider : 1000,
-                plotsPars : plotLabels_2_plotsPars( colorThreadArray ),
                 doPaintGridOnlyOnce : false,
                 doDrawToolline : doDrawToolline(),
 				brightenGrid : 0.3,
@@ -131,28 +130,28 @@
         }
 
         ///horizontal axis x pars, font, etc,
-        function graphAxisX( xColor )
+        function graphAxisX( textColor )
         {
             return {
                 'font-size'     : '18px',
                 fontShiftX      : -12, //in media scale
                 fontShiftY      : +14,
                 decimalDigits   : 3,
-                stroke          : xColor,
-                fill            : xColor,
+                stroke          : textColor,
+                fill            : textColor,
                 'stroke-width'  : '0.2',
             };
         }
 
-        function graphAxisY( yColor )
+        function graphAxisY( textColor )
         {
             return {
                 'font-size'     : '20px',
                 fontShiftX      : -45, //in media scale
                 fontShiftY      : +5,
                 decimalDigits   : 1,
-                stroke          : yColor,
-                fill            : yColor,
+                stroke          : textColor,
+                fill            : textColor,
                 'stroke-width'  : '1',
             };
         }
@@ -215,14 +214,105 @@
         `);
     }
 
+	///this thing is not dynamic (missed in design),
+	///but, colorThreadArray is accessible for reset
+	///dynamically,
+	///
+	//this is just an example how to reset colors dynamically
+	//in model_upcreate():
+	//stdMod.graphFW_lemma.colorThreadArray[0] = sDomF.getFixedColor( 'force' );
+	function setColorThreadArray() {
+		let colorThreadArray = [
+			sDomF.getFixedColor( 'force' ),
+			sDomF.getFixedColor( 'displacement' ),
+		];
+		return colorThreadArray;
+	}
+
+	function setGraphContainerAttributes( digramParentDom$ ) {
+		container$ = $$.div()
+		.addClass( 'chem-equiibr-graph-container' )
+		.to( $$.div().to( digramParentDom$ )
+				.addClass( 'lost-diagram-parent' )
+				//.css( 'position', 'absolute' )
+
+				//:this data sets outer dimensions of the graph
+				.css( 'width', '400px' )
+				.css( 'height', '230px' )
+				.css( 'top', '0' )
+				.css( 'left', '0' )
+				.css( 'z-index', '111111' )
+		);
+		//creates low tire api
+		graph_dimX = 1000;  //innerWidth
+		graph_dimY = 580;   //innerHeight
+		return {container$, graph_dimX, graph_dimY}
+	}
+
+	function setGraphAxes(graphFW, legendText, legendX) {
+		let n2c = sDomF.getFixedColor; //name to color
+		
+		//==================================================
+		// //\\ calls api
+		//==================================================
+		var textColor      = 'rgba(0,0,0,1)';
+		var axisYLegend =
+		[
+			{
+				//together, tobold hover-width and tostroke can be redundant
+				text    :   '<text><tspan class="tofill tobold hover-width"' +
+							'>Force</tspan></text>',
+				x       : 40,
+				y       : 25,
+				style   : {
+							'font-size' : 28 + 'px',
+				},
+			},
+			{	// chart title
+				text    :   '<text><tspan class="tp-force tofill tobold hover-width"' +
+							//overrides tp machinery
+							' style="fill:'+n2c( 'force' ) + '; stroke:'+n2c( 'force' ) + ';"' +
+							'>Actual</tspan>' +
+							'<tspan> and </tspan>' +
+
+							'<tspan class="tp-displacement tofill tobold hover-width"' +
+							//overrides tp machinery
+							' style="fill:'+n2c( 'displacement' ) + '; stroke:' + n2c( 'displacement' ) + ';"' +
+							'>Estimated' +
+							'</tspan>' +
+
+							'<tspan> forces</tspan>' +
+							'</text>',
+				x       : 310,
+				y       : 40,
+				style   : {
+							'font-size' : '30',
+				},
+			},
+		];
+		var axisXLegend =
+		[
+			{
+				text    : legendText,
+				x       : legendX,
+				y       : 25,
+				style   : {
+							'font-size' : '30',
+							'stroke' : textColor,
+							'fill' : textColor,
+				},
+			},
+		];
+		return { textColor, textColor, axisYLegend, axisXLegend, };
+	}
+
 	/**
 	 * Makes a particular graph plot highlight along with its 
 	 * corresponding text.
 	 */
 	///this thing fails if not to synch it with mask,
 	///the unmasked indices must be the same as here:
-	function setsGraphTpClasses(fw)
-	{
+	function setsGraphTpClasses(fw) {
 		fw.plotIx2plotSvg.forEach( (pl,pix) => {
 			switch(pix) {
 				case 0: pl && $$.$(pl).addClass( 'tp-force tostroke' ); break;