+++ changelog/changelog.md
@@ -1,11 +1,14 @@
 # Change Log
 
-## 2026-3-?? (? Version, 0.7)
+## 2026-3-?? (Indigo Version, 0.7)
 ### Added
 * Lemma 6 and beyond
 	* added pinch-to-zoom support
 
 ### Improved
+* Propositions 6 - 13:
+	* minor graph tweaks
+
 * Proposition 13
 	* eccentricity control integrated into the model with point A